Remove 'properties' from DeviceList and DeviceHandleList
Initially, no properties were sent in the DeviceHandleList. After
a property change these were inserted subsequently. This led to
the fact that the roleNames function did not work correctly on
the client side. :
Besides this and for performance reasons it makes no sense to
send the properties of the devices in the lists. "Properties"
has now been removed